CrewPresenters, table on behold page
Use CrewPresenter to display details of selected crew rather than the old CommonCrewData in grid format. Adds a table to make it easier to directly compare numbers of selected crew.

A few things from the old behold format aren't in CrewPresenter:

* Side by side fuse levels
* Build costs
* Event counts
* Hidden traits
* Collections
* Date added (and how originally available)
* More stats (i.e. detailed skill ranks) / player ranks

Some of these could be made available in a CrewPresenter plugin. Opted to show the Big Book writeup below the CrewPresenter, rather than as a plugin; the writeup length is too variable whereas CrewPresenter prefers a more predictable layout. The "More stats" accordion might also be a bad fit for a plugin for the same reason.

The CrewPresenter could be updated to include some of this info more generally. Maybe below the equipment slots, we could add the date added (and how originally available). The portal status could also be grouped with these 2 stats, as they're kind of related.

The table includes some of the missing data, like build costs, event counts, and collection counts.

To do:

* Restore the buff click handler on the CrewPresenter
* Add CrewPresenter to the avatar icons in table view
* Maybe add explanations for certain columns in table view
* Restore permalinks to the page for sharing

Use of PlayerData has been removed from this tool, at least until global player is finalized.