Polestar total odds fix
Restores the proper calculation of the total odds of getting a polestar from owned constellations. Dropped a few nullish coalesces from the calculations in this section to make the confusing math here a little less confusing; the model rightly says the keystones prop is required, so we shouldn't need as many coalesces here.

Updates the player link to the new domain.