Scroll to active player panel
Scrolls to the top of the page when a player panel is requested from the menu.

There's some quirkiness to the scrolling behavior because of the fixed header. Currently the anchor div is styled with a hack to get the scroll to the proper offset; this may not work in mobile, especially if the menu isn't fixed.

One other possible solution is to add this to the stylesheet:
```
* {
   scroll-margin-top: 60px;
}
```