Dashboard, new share profile flow
Introduces a dashboard to the top of datapagelayout to handle non-page-specific components where:
1) Announcements and notifications are displayed
2) Automatic logic is triggered
3) Quick hubs can be activated (perhaps some user settings or maybe a quick search panel)

Player Share is the first working example of a dashboard, handling all the relevant sharing logic, including the share nag, the trigger for automatically updating profile data, notifications about the uploading state, and a panel with the profile link, instructions, and options; for now there's only an auto-update checkbox, but an option to delete public profile should go here too.

Voyage History will get a similar dashboard to automatically reconcile the active voyage with a tracked voyage upon import of player data.

Player Context has a new sessionStates prop to store once-per-session states, like profileUpload (and soon voyageHistoryReconcile). sessionStates are reset to defaults on every player data import.

Note that Player Header remains mostly the same and that the player intake form and related messages are still shown post-pageHeader.