Crew table refinements
Always show stepper when player data is available, even if rosters aren't yet set. This streamlines the rendering and guarantees that the stepper is properly available, even in misaligned roster states. Roster picker should fix any state issues when a different roster is selected.

Loader on rostertable is fixed to the right margin. This works better on desktop, but isn't so great on mobile at the moment. Potential alternatives: loader could be pushed down to crewconfigtable or even searchabletable.

Loader now also listens for changes to tableView.