Send telemetry inline with handleResults
Sends telemetry directly from handleResults, rather than as an effect of the results renderer, which should make this more predictable and less prone to redundant attempts at sending telemetry.

Ideally, telemetry would be triggered manually by the user somehow, rather than automatically as it currently does, because there is no way for the tool to know which of the recommendations (or even which of the requests) is actually going to be used by the user. Right now, the first recommendation of every request is sent to the database.

While I was in there, I isolated the immortal/frozen ratio calculations to telemetry and streamlined the quipment mapping.