Shuttle helper refactor
Major rewrite of the shuttle helper tool, turning it into a standalone tool that works with and without event data (e.g. for help with daily shuttles), as well as with and without player data (e.g. to see the best shuttle crew in the game). The shuttle helper will continue to work as a subcomponent of the event planner.

NOTE: Standalone page is not yet available through navigation, but for the time being can be directly accessed via /shuttlehelper URL.

When 1) there's player data AND 2) there's an active event AND 3) using myCrew as roster, the shuttle helper will import open event shuttle missions and save them across sessions. In all other scenarios, missions and recommendations will NOT be saved. If multiple scenarios are available to a user, they can choose the appropriate mode (standalone only).

New top-down flow instead of the previous tabbed missions/assignments interface. Crew options are moved below the missions table, just before the Recommend Crew button, which also receives greater emphasis here, much like voyage calculator.

Running missions and assignments are now imported with player data and can be used as baseline recommendations.

The missions and assignments tables get a new standardized table component, with better support for sorting both missions and shuttle scores. The tables no longer stack on mobile, but are instead scrollable.

A status column has been added to the missions table, which identifies non-open, open (with expiry times), running, and manually-created missions.

The event projection has been moved to the assignments view and now has options to project multiple scoring scenarios by scheduling shuttles of various durations and regularity. The projection also shows the number of rental tokens and time boosts, when needed.

Various optimizations of the shuttle editor and assignment picker, including a refactor to isolate code better.

NOTE: Most user preferences are NOT saved. This can be added later after functionality of all the options is confirmed.