Adds option to confirm FBB trait solves
Adds an option to Fleet Boss Battles that enables confirmation of solved traits before clearing the node from the UI. This option allows spotters to continue processing nodes when the solution can be deduced manually, while waiting for a fleetmate to actually clear the node in-game. In groups view and the export, the list of crew who can clear that node will continue to be listed until the node is confirmed. The spotter can confirm the solved traits by clicking on the traits or any crew.

Solving nodes by marking crew and solving traits via the traits view behave as usual, bypassing the need for confirmation.

Also adds an undo button when solving (or partially solving) a node via traits.

Spot solves are now marked with a green checkmark circle. "Infallible" solves (i.e. solves authenticated by player data) are now marked with a white checkmark.

Limitations to unconfirmed solves: duplicate counts and coverage are ignored, since the full slate of previous nodes and crew is unknown to the node with the unconfirmed solve. Note that confirmations are also untested in collaboration mode.