Converts voyagers.js to TypeScript
Converts voyagers.js worker to TypeScript.

This is the new baseline for the MVAM voyage calculator. Some of the original code has been streamlined here, but it should otherwise be functionally equivalent.

Expect more refactoring here, both to simplify the code and to better modularize certain parts (i.e. to allow for a drop-in replacement of the lineup assembler, while retaining other structural pieces).