Omega Protocol voyage calc
Experimenting with a new voyage calc, called "Omega Protocol," at least for the time being. This is not yet available from the front end and I would not recommend using it over the existing calcs.

This is an attempt to provide a more exhaustive lineup generator and to see if it produces better results than MVAM's more random approach. Results don't seem to be any better, but there's potential here, especially if we can figure out how to preemptively disqualify a good number of crew combos.