Infinite Diversity
Renamed new voyage calc to "Infinite Diversity," which I think better reflects the philosophy of this new algorithm.

After some failed experiments with cartesian products, I ultimately fell back to the brute force combination approach, albeit with several significant breakthroughs. It actually feels pretty solid now and is close to becoming a viable alternative. In cases where the original calc and/or MVAM yield a better lineup, chances are good IDIC has identified that particular combination of crew too, but it simply disregarded it for some reason. Further investigation of the crew seater and/or estimator may help here.

To facilitate testing and allow direct comparisons with the other calcs, IDIC now appears as an option in the voyage calculator dropdown.