Estimator, sorter changes
The new voyage calc, now called Infinite Diversity, is ready for testing!

Estimator tweaked to work better with both IDIC and MVAM. Scan depth for raised a little for MVAM, but greatly raised for IDIC, which prioritizes thoroughness over speed anyway. Also added a tiebreaker and a special handler for peak-antimatter to the estimator.

New behavior for the sorter. It will return up to 3 of the requested strategy, if the results are tied. If the strategy is "Any best" (an evolution of the now defunct "versatile" strategy), the sorter will return at most 1 lineup per estimate, minimum, and moonshot.

MVAM and IDIC now use the same strategy options, with the following changes: a new "Any best" option is available (see note above) and "Thorough" has been removed. IDIC essentially replaces the thorough option of MVAM.

The peak antimatter strategy does not seem to work with IDIC. Will investigate this later.

Various finetuning of the passing of options between different components.