Localized context
Proposes a rewrite of the translation context, moving as much of the translation handling out of core data into a new localized context.

Localized is exposed on global context, which is the preferred way to access localized strings.

Currently only works with strings from translation json and items json, but can be expanded to handle site strings (like those in CONFIG).

Explicit uses of `traits_named` should be deprecated in favor of translating traits in-place.

Currently no UI for setting a preferred language. When player data loads, the language set in-game will override the preferred language. Should preference override the in-game language? Should site language (site strings) preference be distinct from in-game (game strings) language?