Event model audit
Revises event model, moves voyage event specific details to new eventData prop (activeContent), and moves calculation of crew bonuses for voyage event to crew table (rather than at event planner init).

Note: be careful when using activeContent. The prop only holds details about the active phase of a started event or the first phase of an unstarted event, so it's unreliable for retrieving details about an upcoming phase 2 of a hybrid event.

These changes more consistently limit the eventData model to event-type-agnostic details and more properly anticipate voyage details that may not yet be available (say, when the voyage event is phase 2 of a hybrid).