Peak VP strategy
Introduces a new MVAM strategy option for voyage events: Peak VP. This yields lineups that are higher in starting VP, forcing unfused, low level crew onto the lineup over better voyagers. This typically will result in a lower estimated runtime, but a higher VP rate.

The recommendations now include "projected VP" and "projected VP per minute " for event voyages.

The "dilemma chance" recommendation is now restricted to dilemma voyages.

The "Peak antimatter" strategy has been revised, similar to "Peak VP," in that it will now try to generate a lineup in cases when not every seat trait can be matched. This will still yield the most trait matches for your roster. Because this has a much higher chance of generating a lineup, the experimental tag can be removed.

Fixes a bug in crewfinder to properly restore icons for frozen and active crew.