Refines encounter_traits model
The EventData interface used by the event planner and voyage calc is meant to be event-type and phase agnostic, so the better place for the new encounter_traits prop is on the VoyageEventContent interface. Additionally, other components, particularly the MVAM worker, are expecting voyage event props to be populated on the VoyageEventContent prop.

Moving it out of EventData necessitated moving the parsing of the encounter traits out of getEventData and into its own function in events.ts, guessEncounterTraits, which also consolidates the conversion from the English strings to the trait symbols. If we want to access the encounter_traits in another component, say event planner, that component should similarly use guessEncounterTraits first.

Renames option to requireEncounterTrait; option still disabled pending potential changes to the calc itself.

Moves header of encounter traits inline with the traits.