Voyage calc running ships fix
The voyage calculator now accounts for ships on already running voyages. It will recommend the best non-running ship, except when requesting recommendations from a custom config. In that case, the best ship will be recommended, even if that ship is already running. (Since there's no toggle for considering running ships and I'd rather avoid adding such a toggle with limited practical use, it's easier to simply assume that the best ship is preferred for theoretical recommendations.)