CalcResult model fixes
Fixes CalcResult model and a few references to it.

Note: messageToResults in USSJohnJayHelper should be rewritten to reject a incomplete result rather than assume the result is fully formed (though the result should be complete, which is why I'm okay taking a few shortcuts here).

A fuller remodel of the voyage calculator might also be a good idea, since the interface of a lineup generated from the upcoming lineupeditor doesn't quite match the interface of the lineups generated by the calculators.