Voyage request model
Introduces a new interface to handle voyage calculator requests, with a new type prop to indicate whether the request is for a calculation or an edit (or custom, should we introduce a feature to allow creating a custom lineup from scratch). The Helper class that previously was the request object is now an optional calcHelper prop of the request, which is set only when it's a calculation request.