Enable quipment prospects in lineup editor
Quipment prospects are now properly considered by the lineup editor, dependent on the calculator's current qpConfig preference.

Note: alternate picker crew options are intentionally independent of calculator crew options, so theoretically there should be a front-facing option to toggle quipment prospects for use here without overriding the qpConfig preferences of the calculator crew options. For simplicity, the alternate picker silently respect the calculator qpConfig preferences instead.

Creates a copy of applyQuipmentProspect in new quipment/utils file for use in the lineup editor. The format here anticipates that the config generator/options will be further generalized for use in other components, like the shuttle helper. For the time being, this should be validated and consolidated with the method of the same name in calculator/crewoptions.