Fixes voyageCalcConfig in results
Newly-formed voyageCalcConfigs in the results pane now use crew directly from the proposal. This should address the issue where crew lose their altered properties (like quipment prospects) when passing a proposal from the lineup editor to the results (though why wasn't it doing this in the first place?).

Adds an edit button to a couple of places in the lineup editor summary to make it more obvious that further edits are still possible.