Remodel of voyage calc helper, request
Moves consideredCrew out of the `Helper` interface into `IVoyageRequest` and completely removes voyageConfig and bestShip from `Helper` (those 2 props are already on `Request`) to make it easier and more consistent for certain components to access those props without calculator. Calculations are now initiated by passing in the full `Request`.

A copy of the full roster is now generated by the results group (pre-filtering but post-QP), so that the editor and saved recommendations can use that as consideredCrew. This is also in anticipation of other components that will need full roster.