Adds Featured VP to MVAM
Adds "Featured VP" as a strategy option for MVAM. It follows many of the same rules as "Peak VP," except it prioritizes featured event crew, not just any event crew. Again, this is only for MVAM and has no effect on the original calculator.

Rewords "Any best" strategy to "Any best time" to hopefully help indicate that this particular strategy is only concerned with estimated/minimum/moonshot times (and ignores consideration of peak/featured values).