Enable spot replacement for voyage editor. (#560)
* Enable spot replacement for voyage editor.

* Several new features

Add set default view for people who prefer to see the summary first. (Rememberforever ONLY IF dbid)

Add click on each skill in summary to highlight skill in table.

Add superscript for the number of crew with the specified skill in the table view of the crew picker.

* ADA/508 compliant highlight

* add check to detail.

* enable multiselect AND for highlighted skills.
Code cleanup, bug fixes.
Swap out IconDisplay with AvatarView

* Update datapicker changes to be generic

Instead of passing the replacement, which is very specific to voyage, I have now added the optional parameters "clickAction" and "searchTitle" ... clickAction can override the default 'toggle' effect of single click, and searchTitle can present additional header content above the search bar in search mode.

Also fixed a bug in voyage editor when selecting crew for an empty (or emptied) seat

* fix double title display

* code cleanup

* code cleanup

* Spot replacement overhaul

Summary changes:

The summary page is now the default view and the default view preference has been removed entirely.

A new search button and a note about how to view seat alternates gives users more initial guidance here.

Crew picker changes:

Moves replacement modal header to renderPreface. SemanticUI modal headers can get wonky when they greatly exceed their expected height. The modal sometimes stretches beyond the viewport. The easiest solution is simply to avoid multiline headers. If renderPreface isn't prominent enough, my suggested alternative would be a new optional renderStickyPreface method for DataPicker. (Note that the modal action panel has similar issues; when it exceeds its expected height, some items aren't visible. The solution is the same: keep everything on a single line and optimally don't put too many things there that might render stacked.)

The DataPicker expects a full non-changing dataset, so don't conditionally change the data before sending it. Instead handle everything with filters, passing it to DataPicker as filteredIds. This way the DataPicker knows when to show a note explaining that some results aren't being listed. The options panel additionally explains that "replacement mode" is filtering results for the targeted seat and adds a cancel button to get out of the mode. The cancel button can be moved to the preface. One other suggestion would be to add a button in the note (or preface) to return to the summary view, so you can pick a different seat.

Memoizes tableSetup so that the sort override doesn't override the sort on every render.

Slot picker changes:

The selected crew from the crew picker completely bypasses the slot picker. Estimates are the most expensive operation here, so if we're not using the slot picker for replacements, might as well do the single estimate and go straight back to the summary.

Skill check accordion:

Disables skill highlighting on the results skill check. Skill highlighting can only be done with the skill check of the lineup editor summary.

* add highlight hook for skill check

* replace code with hook

* Enable highlighting in main view with skill check.

This involved memoizing myCrew, also, I simplified the getSkills function.

* make sure function returns clean skills.

* Implement highlight in calculator, too.

* Minor bug fixes

Slot replacement properly removes a past assignment to a crew moving to a new seat.

Sort by antimatter on slot picker fixed.

Prospect mode is now responsive to the reset button on crew picker options.

* remove grade coloring from skill number hint

---------

Co-authored-by: ussjohnjay <jtjasmin+stt@gmail.com>