DataGrid default items per row
Changes the default number of items per row in a DataGrid  presenter (i.e. desktop: 4, tablet: 3, and mobile: 2), which is more ideal for the typical use case (i.e. a crew picker modal). Use the new columnProps field to change this behavior, say, if the grid column is expected to be wider than a typical crew avatar.