Encounter helper QOL
Re-enables encounter helper. Note that it is now listed after the rewards accordion.

Shows new info card for imported encounters, which provides an at-a-glance overview of the encounter detailing contest skills, crit traits, and VP multiplier. This card also shows a message when it detects an in-progress encounter, explaining that the tool can't see previously used crew.

Adds visual cues for encounters with a boss.

Adds option to filter champions table by skill. Targeted skills will be highlighted in the column headers. Tapping a row in the contests table will automatically select that contest's skills. This is a variation on a requested feature and remains a work-in-progress.

Adds a visual cue to champions table when a crew is assigned to a contest.

Tweaks encounter importer to handle encounter data (i.e. VoyageRefreshEncounter) separately from serialized encounter object (IEncounter).

Adds/tweaks various descriptive text.