Add crew excluder to shuttle helper. Experimental prototype.  Gave the crew excluder some autonomy from the voyage context.
Untangled it somewhat from the voyage calculator by removing dependencies on the calculatorContext, adding voyageConfig and events as optional props, and  copied the getEvents() function over to CrewExcluder and running it conditionally.  CrewExcluder can now pull event data independently, and be used virtually anywhere.

This is potentially not the final form for this, but I wanted to get the ball rolling in order