Make DataContext autosync aware. Autosync now publishes sync_timestamp.txt to /static/structured every time it is run.
DataContext will now load this file fresh from the remote every time and compare it to the saved value before allowing any demand fulfillment.

If the timestamp has changed, then a new key is issued and wantFresh is sent.  The key is appended to the url as a query string parameter, forcing the browser to require a fresh copy.