Model untangling and reorganization Moving types into different files to reflect their relationship to other models and methods in the app.  Some of them were clustered in illogical and far-out files that made no sense.  So, I moved them. There were many references.
Ran npx tsc --noEmit to confirm code sanity.  All passed, no errors in linting or putting hands on objects.

This follows on from a patch I pushed to master to correct some of the gauntlet crit calculation patterns.

This code is also shared with DataScore.  There will be a score correction forthcoming.